backend/pyproject.toml
[tool.poetry]
name = "casapp-backend"
version = "0.1.0"
description = "Real estate asset management backend"
authors = ["Francesco Vigni <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.3"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
dj-database-url = "^2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.8"
pytest-cov = "^5.0"
model-bakery = "^1.17"
black = "^24.8"
ruff = "^0.5.6"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=apps --cov-report=term-missing --cov-report=xml --cov-fail-under=100"

[tool.coverage.run]
branch = true
source = ["apps"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
